refactor: reduce code duplication between add-menu-item-button and update-menu-item-button
