+++ manager-application/src/app/item-dialog/error-dialog.html
@@ -1,0 +1,5 @@
+<h2 mat-dialog-title>An error happened!</h2>
+<mat-dialog-content>
+  <h3>It was impossible to complete the operation!</h3>
+  <p>Error: {{errorMessage}}</p>
+</mat-dialog-content>
