+++ manager-application/src/app/item-dialog/item-dialog.component.css
@@ -1,0 +1,28 @@
+.button{
+    width: auto;
+    height: auto;
+    padding: 2%;
+}
+
+button{
+    background-color: #ff4081 !important;
+}
+
+.add-button{
+    width: 100%;
+}
+
+.qty{
+    width: 30%;
+    margin: 5%;
+}
+
+.error{
+    color: red;
+    font-weight: bold;
+}
+
+h2{
+    margin-left: 5%;
+    margin-top: 5%;
+}
