+++ manager-application/src/utils/DialogData.ts
@@ -1,4 +1,5 @@
 import { MatDialog } from "@angular/material/dialog";
+import { Item } from "./Item";
 
 /**
  * This interface represents the data passed to a dialog
@@ -6,4 +7,8 @@ import { MatDialog } from "@angular/material/dialog";
 export interface DialogData {
 	ws: WebSocket,
 	dialog: MatDialog
+	update: boolean,
+	title: string,
+	buttonMsg: string,
+	item?: Item
 }
